Fix round-3 CI errors: LP64 uint32, source inclusion, lcov compat
- Fix LP64 data model: uint32 typedef from unsigned long to unsigned int
  across Platform_Types.h, sc_types.h, and 60+ test files (8-byte long
  on 64-bit Linux caused seed-key mismatches in TCU security tests)
- Convert CVC test_Ssd1306 and FZC test_Swc_Brake to source inclusion
  pattern with header guards (fixes linker undefined reference errors)
- Remove unused lcov exclude pattern '*/unity/*' in integration Makefile
  (lcov 2.x treats unused patterns as errors, exit code 25)
- Fix BSW mock_setsockopt unused function warning (-Werror)
- Remove unused 'raw' variable in RZC test_Swc_RzcNvm

+++ firmware/bcm/test/test_Swc_BcmCan.c
@@ -20,8 +20,8 @@
 
 typedef unsigned char  uint8;
 typedef unsigned short uint16;
-typedef unsigned long  uint32;
-typedef signed long    sint32;
+typedef unsigned int  uint32;
+typedef signed int    sint32;
 typedef uint8          Std_ReturnType;
 typedef uint8          boolean;
